feat(v2): render Markdown on server-side instead
