+++ v2/lib/webpack/loaders/toSlug.js
